Remove tag retrieval from Playlist
